Fixed parameter in json messages

+++ app/converters/messages/Json/Item.scala
@@ -3,6 +3,6 @@ package converters.messages.Json
 case class Item(
     description: String,
     quantity: Int,
-    pricePerItem: Int,
+    pricePerItem: Float,
     haveVat: Boolean,
     vatRate: Double)
